refactor: move pagination to a new file
