+++ src/helpers/print_devlog_table.rs
@@ -17,7 +17,7 @@
 
 use crate::commands::devlog::list::DevlogFields;
 use crate::models::devlog::Devlog;
-use crate::models::devlog_vec::Pagination;
+use crate::models::pagination::Pagination;
 use chrono::{DateTime, Local};
 use comfy_table::modifiers::UTF8_ROUND_CORNERS;
 use comfy_table::presets::UTF8_FULL;
