+++ src/models/devlog_vec.rs
@@ -16,15 +16,8 @@
 // along with flavorcli.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::models::devlog::Devlog;
+use crate::models::pagination::Pagination;
 use serde::Deserialize;
-
-#[derive(Debug, Deserialize)]
-pub struct Pagination {
-    pub current_page: u32,
-    pub total_pages: u32,
-    pub total_count: u32,
-    pub next_page: Option<u32>,
-}
 
 #[derive(Debug, Deserialize)]
 pub struct DevlogVec {
