+++ src/models/mod.rs
@@ -23,3 +23,4 @@ pub mod project_vec;
 pub mod store;
 pub mod user;
 pub mod user_vec;
+pub mod pagination;
